distlock/postgres: convert key to int64
The actual datatype expected by the postgres function is an int64, and
the wrapper doesn't like a caller trying to use a value not
representable in that range.

+++ pkg/distlock/postgres/distlock.go
@@ -107,7 +107,7 @@ func (l *lock) TryLock(ctx context.Context, key string) (bool, error) {
 
 	kh := fnv.New64a()
 	_, _ = fmt.Fprint(kh, key) // Writing to a hash.Hash never errors.
-	keyInt64 := kh.Sum64()
+	keyInt64 := int64(kh.Sum64())
 
 	// start transaction
 	tx, err := l.db.Beginx()
